[GEOT-7345] WFS Capabilities from local file
